Alternative readme for installing the mod
This is a method that uses xcopy instead of mklink. It is the only reliable way to setup this mod that I know off. I've been helping another developer try to setup his local version, and he is at wits end and is about to give up. This will hopefully help.