Possibly fix reincarnation and meepo bug?
I played around with this code, and it seems to have fixed the meepo bug and the recincarnation bug. I cant be sure unless its test ingame. I'm posting this issue just as a suggestion, I expect you to reject the merge, but hopefully it helps you find the bug.